app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-community==0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pypdf>=3.17.4
google-generativeai>=0.3.1
python-multipart>=0.0.6
